Added State Pattern For Screens
Added WheelScreen,RewardScreen

+++ Assets/Scripts/ScriptableObject/WheelSO.cs
@@ -10,12 +10,12 @@ public class WheelSO : ScriptableObject
     public e_Spin_Direction e_Spin_Direction;
     public Sprite baseSprite, indicatorSprite;
     public AnimationCurve spinCurve;
-    public float minSpinTime = 3, maxSpinTime = 4;
+    public int minSpinTime = 3, maxSpinTime = 4;
     public Sprite GetWheelBaseSprite() { return baseSprite; }
     public Sprite GetWheelIndicatorSprite() { return indicatorSprite; }
-    public float GetRandomSpintTime()
+    public int GetRandomSpintTime()
     {
-        float randomValue = UnityEngine.Random.Range(minSpinTime, maxSpinTime);
+        int randomValue = UnityEngine.Random.Range(minSpinTime, maxSpinTime);
         return randomValue;
     }
 }
